src/vertex-agent-adk/pyproject.toml
[project]
name = "vertex-agent-adk"
version = "0.1.0"
description = "ADK to Agent Engine Test"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform[agent_engines,adk]>=1.140.0",
    "cloudpickle>=3.1.2",
    "fire>=0.7.1",
    "google-cloud-storage",
    "redis",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agente_cortes", "client"]

[dependency-groups]
dev = []
